backend/requirements.txt
Django==4.2.2
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
python-dotenv==1.0.0